# backend/openapi.yaml
openapi: 3.0.1
info:
  title: AguaVigia CTG — API publica
  description: |-
    Monitoreo ciudadano del acueducto de Cartagena de Indias.

    Plataforma ciudadana e independiente: no esta afiliada a Aguas de Cartagena
    S.A. E.S.P. ni a ninguna entidad distrital.
  contact:
    name: AguaVigia CTG
  license:
    name: Proyecto de aula
  version: 1.0.0
servers:
- url: /
  description: Mismo origen que sirve la API
...
paths:
...
  /api/veedor/cortes:
    get:
      tags:
      - Veedor - Cortes
      summary: Listar los cortes que afectan a un sector
      operationId: listarPorSector
      parameters:
      - name: sectorId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorteRespuesta"
        "401":
          description: "Sin sesión, token inválido, caducado o revocado"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: La sesión no tiene el permiso que exige esta operación
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    post:
      tags:
      - Veedor - Cortes
      summary: Registrar un corte oficial
      description: "Sectores afectados, inicio, fin prometido y causa (RF016). Origen\
        \ VEEDOR."
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCorte"
        required: true
      responses:
        "201":
          description: Corte registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "400":
          description: Datos inválidos o algún sector no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Sin sesión, token inválido, caducado o revocado"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: La sesión no tiene el permiso que exige esta operación
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/reportes:
    post:
      tags:
      - Reportes
      summary: Registrar un reporte ciudadano
      description: |-
        Sin registro ni cuenta (RF005). Limita automáticamente los reportes por
        dispositivo en la ventana vigente (RF006) — ver 429. Hace falta el `sectorId`, la
        `coordenada` o ambos (RF007): con solo la coordenada el servidor infiere el sector
        que la contiene y responde 400 si cae fuera de todo barrio de Cartagena. La
        coordenada se envía solo si el usuario autorizó compartir su ubicación.
      operationId: registrar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudReporte"
        required: true
      responses:
        "201":
          description: Reporte registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "400":
          description: "Sector inexistente, tipo inválido, huella fuera de 32-128\
            \ caracteres, coordenada fuera de Cartagena o sin sector ni coordenada"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: El dispositivo superó el límite de reportes para este sector
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/veedor/cortes/{id}/cierre:
    patch:
      tags:
      - Veedor - Cortes
      summary: Cerrar un corte con la hora real de restablecimiento (RF017)
      operationId: cerrar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCierreCorte"
        required: true
      responses:
        "200":
          description: Corte cerrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "404":
          description: El corte no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: El corte ya estaba cerrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Sin sesión, token inválido, caducado o revocado"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: La sesión no tiene el permiso que exige esta operación
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/veedor/cortes/{id}:
    get:
      tags:
      - Veedor - Cortes
      summary: Consultar un corte por su identificador
      operationId: consultar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Corte encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "404":
          description: No existe un corte con ese id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Sin sesión, token inválido, caducado o revocado"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: La sesión no tiene el permiso que exige esta operación
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/cumplimiento/cortes/{corteId}:
    get:
      tags:
      - Cumplimiento
      summary: Índice de un corte cerrado
      operationId: porCorte
      parameters:
      - name: corteId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Índice calculado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndiceCumplimientoRespuesta"
        "404":
          description: El corte no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: El corte todavía no está cerrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    SolicitudCorte:
      required:
      - finPrometido
      - inicio
      type: object
      properties:
        sectoresAfectados:
          minItems: 1
          type: array
          description: Identificadores de los sectores afectados
          example:
          - manga
          - bocagrande
          items:
            type: string
            description: Identificadores de los sectores afectados
            example: "[\"manga\",\"bocagrande\"]"
        inicio:
          type: string
          format: date-time
        finPrometido:
          type: string
          format: date-time
        causa:
          minLength: 1
          type: string
          example: Mantenimiento planta El Bosque
      description: Registro de un corte oficial por el veedor (RF016)
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    CorteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectoresAfectados:
          type: array
          items:
            type: string
        inicio:
          type: string
          format: date-time
        finPrometido:
          type: string
          format: date-time
        finReal:
          type: string
          description: Nulo mientras el corte sigue abierto
          format: date-time
        causa:
          type: string
        origen:
          type: string
        estado:
          type: string
      description: Corte oficial (RF016-RF017)
...
    CoordenadaDTO:
      required:
      - latitud
      - longitud
      type: object
      properties:
        latitud:
          maximum: 90
          exclusiveMaximum: false
          minimum: -90
          exclusiveMinimum: false
          type: number
          format: double
        longitud:
          maximum: 180
          exclusiveMaximum: false
          minimum: -180
          exclusiveMinimum: false
          type: number
          format: double
      description: "Coordenada GPS del reporte, solo cuando el usuario la autoriza\
        \ (RF007)"
    SolicitudReporte:
      type: object
      properties:
        sectorId:
          type: string
          description: |-
            Identificador del sector reportado. Opcional si viaja la coordenada: entonces el
            servidor infiere el barrio que la contiene (RF007). Si no viaja ninguno, 400.
          nullable: true
          example: bocagrande
        tipo:
          minLength: 1
          type: string
          description: "SIN_AGUA, PRESION_BAJA o SERVICIO_RESTABLECIDO"
          example: SIN_AGUA
        huella:
          maxLength: 128
          minLength: 32
          type: string
          description: |-
            Huella anónima del dispositivo (ADR-007) — no es una cuenta ni un identificador
            personal. El cliente la genera una vez (p. ej. un UUID persistido en el dispositivo,
            hasheado) y la reutiliza en cada reporte; es lo único que permite RF006 (límite de
            reportes por dispositivo) sin pedir registro.
        coordenada:
          $ref: "#/components/schemas/CoordenadaDTO"
      description: Reporte ciudadano sin registro (RF005-RF008)
    ReporteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectorId:
          type: string
        tipo:
          type: string
        timestamp:
          type: string
          format: date-time
        fotoUrl:
          type: string
        confirmaciones:
          type: integer
          format: int32
      description: Reporte ciudadano registrado
...
    SolicitudCierreCorte:
      required:
      - horaReal
      type: object
      properties:
        horaReal:
          type: string
          format: date-time
      description: Cierre de un corte con la hora real de restablecimiento (RF017)
...
    IndiceCumplimientoRespuesta:
      type: object
      properties:
        sectorId:
          type: string
          description: "Nulo cuando el índice es por corte o global, no por sector"
        duracionPrometidaSegundos:
          type: integer
          format: int64
        duracionRealSegundos:
          type: integer
          format: int64
        desviacionSegundos:
          type: integer
          description: duracionReal - duracionPrometida. Negativa si terminó antes
            de lo prometido
          format: int64
        porcentajeCumplimiento:
          type: number
          description: Capado en 100 cuando el corte termina antes o a tiempo
          format: double
      description: |-
        Índice de Cumplimiento (RF020-RF022): comparación explícita entre duración prometida y
        real, nunca un porcentaje aislado.
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |-
        Token que devuelve POST /api/veedor/sesion, en la cabecera
        `Authorization: Bearer <token>`. Caduca a las 8 horas y no se renueva: al
        caducar hay que volver a iniciar sesión.
      scheme: bearer
      bearerFormat: JWT